make_index/requirements.txt
unidecode
psycopg2-binary
Flask
tabulate
alphabet-detector
lb-matching-tools@git+https://github.com/metabrainz/listenbrainz-matching-tools.git@v2024.01.30.1
peewee
tqdm
scikit-learn
nmslib-metabrainz
